LibraryApp/requirements.txt
pyinstaller
pandas
openpyxl
python-docx
tkcalendar